Share extension methods instead of duplicating them
