--- src/CodeStyle/VisualBasic/Analyzers/Formatting/SyntaxTriviaExtensions.vb
@@ -1,28 +1,0 @@
-﻿' Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-Imports System.Runtime.CompilerServices
-
-Namespace Microsoft.CodeAnalysis.VisualBasic.Extensions
-    Friend Module SyntaxTriviaExtensions
-        <Extension()>
-        Public Function IsKind(trivia As SyntaxTrivia, kind1 As SyntaxKind, kind2 As SyntaxKind) As Boolean
-            Return trivia.Kind = kind1 OrElse
-                   trivia.Kind = kind2
-        End Function
-
-        <Extension>
-        Public Function IsWhitespaceOrEndOfLine(trivia As SyntaxTrivia) As Boolean
-            Return trivia.IsWhitespace() OrElse trivia.IsEndOfLine()
-        End Function
-
-        <Extension>
-        Public Function IsWhitespace(trivia As SyntaxTrivia) As Boolean
-            Return trivia.Kind = SyntaxKind.WhitespaceTrivia
-        End Function
-
-        <Extension>
-        Public Function IsEndOfLine(trivia As SyntaxTrivia) As Boolean
-            Return trivia.Kind = SyntaxKind.EndOfLineTrivia
-        End Function
-    End Module
-End Namespace
