+++ src/Workspaces/CSharp/Portable/Extensions/MemberDeclarationSyntaxExtensions_GetAttributes.cs
