+++ src/Workspaces/CSharp/Portable/Extensions/SyntaxNodeExtensions.cs
@@ -17,105 +17,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Extensions
 {
     internal static partial class SyntaxNodeExtensions
     {
-        public static bool IsKind<TNode>(this SyntaxNode node, SyntaxKind kind, out TNode result)
-            where TNode : SyntaxNode
-        {
-            if (node.IsKind(kind))
-            {
-                result = (TNode)node;
-                return true;
-            }
-
-            result = null;
-            return false;
-        }
-
-        public static bool IsParentKind(this SyntaxNode node, SyntaxKind kind)
-        {
-            return node != null && CodeAnalysis.CSharpExtensions.IsKind(node.Parent, kind);
-        }
-
-        public static bool IsParentKind(this SyntaxNode node, SyntaxKind kind1, SyntaxKind kind2)
-        {
-            return node != null && IsKind(node.Parent, kind1, kind2);
-        }
-
-        public static bool IsParentKind(this SyntaxNode node, SyntaxKind kind1, SyntaxKind kind2, SyntaxKind kind3)
-        {
-            return node != null && IsKind(node.Parent, kind1, kind2, kind3);
-        }
-
-        public static bool IsParentKind(this SyntaxNode node, SyntaxKind kind1, SyntaxKind kind2, SyntaxKind kind3, SyntaxKind kind4)
-        {
-            return node != null && IsKind(node.Parent, kind1, kind2, kind3, kind4);
-        }
-
-        public static bool IsKind(this SyntaxNode node, SyntaxKind kind1, SyntaxKind kind2)
-        {
-            if (node == null)
-            {
-                return false;
-            }
-
-            var csharpKind = node.Kind();
-            return csharpKind == kind1 || csharpKind == kind2;
-        }
-
-        public static bool IsKind(this SyntaxNode node, SyntaxKind kind1, SyntaxKind kind2, SyntaxKind kind3)
-        {
-            if (node == null)
-            {
-                return false;
-            }
-
-            var csharpKind = node.Kind();
-            return csharpKind == kind1 || csharpKind == kind2 || csharpKind == kind3;
-        }
-
-        public static bool IsKind(this SyntaxNode node, SyntaxKind kind1, SyntaxKind kind2, SyntaxKind kind3, SyntaxKind kind4)
-        {
-            if (node == null)
-            {
-                return false;
-            }
-
-            var csharpKind = node.Kind();
-            return csharpKind == kind1 || csharpKind == kind2 || csharpKind == kind3 || csharpKind == kind4;
-        }
-
-        public static bool IsKind(this SyntaxNode node, SyntaxKind kind1, SyntaxKind kind2, SyntaxKind kind3, SyntaxKind kind4, SyntaxKind kind5)
-        {
-            if (node == null)
-            {
-                return false;
-            }
-
-            var csharpKind = node.Kind();
-            return csharpKind == kind1 || csharpKind == kind2 || csharpKind == kind3 || csharpKind == kind4 || csharpKind == kind5;
-        }
-
-        public static bool IsKind(this SyntaxNode node, SyntaxKind kind1, SyntaxKind kind2, SyntaxKind kind3, SyntaxKind kind4, SyntaxKind kind5, SyntaxKind kind6)
-        {
-            if (node == null)
-            {
-                return false;
-            }
-
-            var csharpKind = node.Kind();
-            return csharpKind == kind1 || csharpKind == kind2 || csharpKind == kind3 || csharpKind == kind4 || csharpKind == kind5 || csharpKind == kind6;
-        }
-
-        public static bool IsKind(this SyntaxNode node, SyntaxKind kind1, SyntaxKind kind2, SyntaxKind kind3, SyntaxKind kind4, SyntaxKind kind5, SyntaxKind kind6, SyntaxKind kind7, SyntaxKind kind8, SyntaxKind kind9, SyntaxKind kind10, SyntaxKind kind11)
-        {
-            if (node == null)
-            {
-                return false;
-            }
-
-            var csharpKind = node.Kind();
-            return csharpKind == kind1 || csharpKind == kind2 || csharpKind == kind3 || csharpKind == kind4 || csharpKind == kind5 || csharpKind == kind6 || csharpKind == kind7 || csharpKind == kind8 || csharpKind == kind9 || csharpKind == kind10 || csharpKind == kind11;
-        }
-
         /// <summary>
         /// Returns the list of using directives that affect <paramref name="node"/>. The list will be returned in
         /// top down order.  
@@ -215,35 +116,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Extensions
             }
         }
 
-        public static IEnumerable<SyntaxTrivia> GetAllPrecedingTriviaToPreviousToken(
-            this SyntaxNode node, SourceText sourceText = null,
-            bool includePreviousTokenTrailingTriviaOnlyIfOnSameLine = false)
-            => node.GetFirstToken().GetAllPrecedingTriviaToPreviousToken(
-                sourceText, includePreviousTokenTrailingTriviaOnlyIfOnSameLine);
-
-        /// <summary>
-        /// Returns all of the trivia to the left of this token up to the previous token (concatenates
-        /// the previous token's trailing trivia and this token's leading trivia).
-        /// </summary>
-        public static IEnumerable<SyntaxTrivia> GetAllPrecedingTriviaToPreviousToken(
-            this SyntaxToken token, SourceText sourceText = null,
-            bool includePreviousTokenTrailingTriviaOnlyIfOnSameLine = false)
-        {
-            var prevToken = token.GetPreviousToken(includeSkipped: true);
-            if (prevToken.Kind() == SyntaxKind.None)
-            {
-                return token.LeadingTrivia;
-            }
-
-            if (includePreviousTokenTrailingTriviaOnlyIfOnSameLine &&
-                !sourceText.AreOnSameLine(prevToken, token))
-            {
-                return token.LeadingTrivia;
-            }
-
-            return prevToken.TrailingTrivia.Concat(token.LeadingTrivia);
-        }
-
         public static bool IsBreakableConstruct(this SyntaxNode node)
         {
             switch (node.Kind())
@@ -330,14 +202,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Extensions
             return node.IsKind(SyntaxKind.MethodDeclaration)
                 || node.IsAnyLambdaOrAnonymousMethod()
                 || node.IsKind(SyntaxKind.LocalFunctionStatement);
-        }
-
-        public static bool IsAnyArgumentList(this SyntaxNode node)
-        {
-            return node.IsKind(SyntaxKind.ArgumentList) ||
-                   node.IsKind(SyntaxKind.AttributeArgumentList) ||
-                   node.IsKind(SyntaxKind.BracketedArgumentList) ||
-                   node.IsKind(SyntaxKind.TypeArgumentList);
         }
 
         public static bool IsAnyLambda(this SyntaxNode node)
@@ -666,29 +530,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Extensions
             return node != null ? node.Parent : null;
         }
 
-        public static (SyntaxToken openBrace, SyntaxToken closeBrace) GetBraces(this SyntaxNode node)
-        {
-            switch (node)
-            {
-                case NamespaceDeclarationSyntax namespaceNode:
-                    return (namespaceNode.OpenBraceToken, namespaceNode.CloseBraceToken);
-                case BaseTypeDeclarationSyntax baseTypeNode:
-                    return (baseTypeNode.OpenBraceToken, baseTypeNode.CloseBraceToken);
-                case AccessorListSyntax accessorListNode:
-                    return (accessorListNode.OpenBraceToken, accessorListNode.CloseBraceToken);
-                case BlockSyntax blockNode:
-                    return (blockNode.OpenBraceToken, blockNode.CloseBraceToken);
-                case SwitchStatementSyntax switchStatementNode:
-                    return (switchStatementNode.OpenBraceToken, switchStatementNode.CloseBraceToken);
-                case AnonymousObjectCreationExpressionSyntax anonymousObjectCreationExpression:
-                    return (anonymousObjectCreationExpression.OpenBraceToken, anonymousObjectCreationExpression.CloseBraceToken);
-                case InitializerExpressionSyntax initializeExpressionNode:
-                    return (initializeExpressionNode.OpenBraceToken, initializeExpressionNode.CloseBraceToken);
-            }
-
-            return default;
-        }
-
         public static (SyntaxToken openBrace, SyntaxToken closeBrace) GetParentheses(this SyntaxNode node)
         {
             switch (node)
@@ -730,39 +571,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Extensions
                 case AttributeListSyntax n: return (n.OpenBracketToken, n.CloseBracketToken);
                 case BracketedParameterListSyntax n: return (n.OpenBracketToken, n.CloseBracketToken);
                 default: return default;
-            }
-        }
-
-        public static bool IsEmbeddedStatementOwner(this SyntaxNode node)
-        {
-            return
-                   node is DoStatementSyntax ||
-                   node is ElseClauseSyntax ||
-                   node is FixedStatementSyntax ||
-                   node is CommonForEachStatementSyntax ||
-                   node is ForStatementSyntax ||
-                   node is IfStatementSyntax ||
-                   node is LabeledStatementSyntax ||
-                   node is LockStatementSyntax ||
-                   node is UsingStatementSyntax ||
-                   node is WhileStatementSyntax;
-        }
-
-        public static StatementSyntax GetEmbeddedStatement(this SyntaxNode node)
-        {
-            switch (node)
-            {
-                case DoStatementSyntax n: return n.Statement;
-                case ElseClauseSyntax n: return n.Statement;
-                case FixedStatementSyntax n: return n.Statement;
-                case CommonForEachStatementSyntax n: return n.Statement;
-                case ForStatementSyntax n: return n.Statement;
-                case IfStatementSyntax n: return n.Statement;
-                case LabeledStatementSyntax n: return n.Statement;
-                case LockStatementSyntax n: return n.Statement;
-                case UsingStatementSyntax n: return n.Statement;
-                case WhileStatementSyntax n: return n.Statement;
-                default: return null;
             }
         }
