+++ src/Workspaces/CSharp/Portable/Extensions/SyntaxTokenExtensions_SharedWithCodeStyle.cs
@@ -1,12 +1,10 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
 using System.Linq;
-using Microsoft.CodeAnalysis.Shared.Extensions;
-using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.CSharp.Extensions
 {
-    internal static class SyntaxTokenExtensions
+    internal static partial class SyntaxTokenExtensions
     {
         public static bool IsKindOrHasMatchingText(this SyntaxToken token, SyntaxKind kind)
         {
@@ -16,6 +14,36 @@ namespace Microsoft.CodeAnalysis.CSharp.Extensions
         public static bool HasMatchingText(this SyntaxToken token, SyntaxKind kind)
         {
             return token.ToString() == SyntaxFacts.GetText(kind);
+        }
+
+        public static bool IsKind(this SyntaxToken token, SyntaxKind kind1, SyntaxKind kind2)
+        {
+            return token.Kind() == kind1
+                || token.Kind() == kind2;
+        }
+
+        public static bool IsKind(this SyntaxToken token, SyntaxKind kind1, SyntaxKind kind2, SyntaxKind kind3)
+        {
+            return token.Kind() == kind1
+                || token.Kind() == kind2
+                || token.Kind() == kind3;
+        }
+
+        public static bool IsKind(this SyntaxToken token, SyntaxKind kind1, SyntaxKind kind2, SyntaxKind kind3, SyntaxKind kind4)
+        {
+            return token.Kind() == kind1
+                || token.Kind() == kind2
+                || token.Kind() == kind3
+                || token.Kind() == kind4;
+        }
+
+        public static bool IsKind(this SyntaxToken token, SyntaxKind kind1, SyntaxKind kind2, SyntaxKind kind3, SyntaxKind kind4, SyntaxKind kind5)
+        {
+            return token.Kind() == kind1
+                || token.Kind() == kind2
+                || token.Kind() == kind3
+                || token.Kind() == kind4
+                || token.Kind() == kind5;
         }
 
         public static bool IsKind(this SyntaxToken token, params SyntaxKind[] kinds)
@@ -32,13 +60,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Extensions
         public static bool IsOpenBraceOfAccessorList(this SyntaxToken token)
         {
             return token.IsKind(SyntaxKind.OpenBraceToken) && token.Parent.IsKind(SyntaxKind.AccessorList);
-        }
-
-        public static bool IsLastTokenOfNode<T>(this SyntaxToken token)
-            where T : SyntaxNode
-        {
-            var node = token.GetAncestor<T>();
-            return node != null && token == node.GetLastToken(includeZeroWidth: true);
         }
     }
 }
