+++ src/Workspaces/Core/Portable/Shared/Extensions/SourceTextExtensions.cs
@@ -12,7 +12,7 @@ using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.Shared.Extensions
 {
-    internal static class SourceTextExtensions
+    internal static partial class SourceTextExtensions
     {
         /// <summary>
         /// Returns the leading whitespace of the line located at the specified position in the given snapshot.
@@ -50,57 +50,6 @@ namespace Microsoft.CodeAnalysis.Shared.Extensions
         {
             text.GetLineAndOffset(textSpan.Start, out startLineNumber, out startOffset);
             text.GetLineAndOffset(textSpan.End, out endLineNumber, out endOffset);
-        }
-
-        public static bool OverlapsHiddenPosition(
-            this SourceText text, TextSpan span, Func<int, CancellationToken, bool> isPositionHidden, CancellationToken cancellationToken)
-        {
-            var result = TryOverlapsHiddenPosition(text, span, isPositionHidden, cancellationToken);
-            cancellationToken.ThrowIfCancellationRequested();
-            return result;
-        }
-
-        /// <summary>
-        /// Same as OverlapsHiddenPosition but doesn't throw on cancellation.  Instead, returns false
-        /// in that case.
-        /// </summary>
-        public static bool TryOverlapsHiddenPosition(
-            this SourceText text, TextSpan span, Func<int, CancellationToken, bool> isPositionHidden,
-            CancellationToken cancellationToken)
-        {
-            var startLineNumber = text.Lines.IndexOf(span.Start);
-            var endLineNumber = text.Lines.IndexOf(span.End);
-
-            // NOTE(cyrusn): It's safe to examine the start of a line because you can't have a line
-            // with both a pp directive and code on it.  so, for example, if a node crosses a region
-            // then it must be the case that the start of some line from the start of the node to
-            // the end is hidden.  i.e.:
-#if false
-'           class C
-'           {
-'#line hidden
-'           }
-'#line default
-#endif
-            // The start of the line with the } on it is hidden, and thus the node overlaps a hidden
-            // region.
-
-            for (var lineNumber = startLineNumber; lineNumber <= endLineNumber; lineNumber++)
-            {
-                if (cancellationToken.IsCancellationRequested)
-                {
-                    break;
-                }
-
-                var linePosition = text.Lines[lineNumber].Start;
-                var isHidden = isPositionHidden(linePosition, cancellationToken);
-                if (isHidden)
-                {
-                    return true;
-                }
-            }
-
-            return false;
         }
 
         public static TextChangeRange GetEncompassingTextChangeRange(this SourceText newText, SourceText oldText)
@@ -187,13 +136,6 @@ namespace Microsoft.CodeAnalysis.Shared.Extensions
         private static bool Match(char normalizedLeft, char right, bool caseSensitive)
         {
             return caseSensitive ? normalizedLeft == right : normalizedLeft == CaseInsensitiveComparison.ToLower(right);
-        }
-
-        public static bool AreOnSameLine(this SourceText text, SyntaxToken token1, SyntaxToken token2)
-        {
-            return token1.RawKind != 0 &&
-                token2.RawKind != 0 &&
-                text.Lines.IndexOf(token1.Span.End) == text.Lines.IndexOf(token2.SpanStart);
         }
 
         // 32KB. comes from SourceText char buffer size and less than large object size
