+++ src/Workspaces/Core/Portable/Shared/Extensions/SyntaxTreeExtensions_SharedWithCodeStyle.cs
@@ -5,7 +5,7 @@ using Microsoft.CodeAnalysis.Text;
 
 namespace Microsoft.CodeAnalysis.Shared.Extensions
 {
-    internal static class SyntaxTreeExtensions
+    internal static partial class SyntaxTreeExtensions
     {
         public static bool OverlapsHiddenPosition(this SyntaxTree tree, TextSpan span, CancellationToken cancellationToken)
         {
