+++ src/Workspaces/VisualBasic/Portable/Extensions/SyntaxTokenExtensions_SharedWithCodeStyle.vb
@@ -3,11 +3,16 @@
 Imports System.Runtime.CompilerServices
 
 Namespace Microsoft.CodeAnalysis.VisualBasic.Extensions
-    Friend Module SyntaxTokenExtensions
+    Partial Friend Module SyntaxTokenExtensions
         <Extension()>
         Public Function IsKind(token As SyntaxToken, kind1 As SyntaxKind, kind2 As SyntaxKind) As Boolean
             Return token.Kind = kind1 OrElse
                    token.Kind = kind2
+        End Function
+
+        <Extension()>
+        Public Function IsKind(token As SyntaxToken, ParamArray kinds As SyntaxKind()) As Boolean
+            Return kinds.Contains(token.Kind)
         End Function
     End Module
 End Namespace
